Shaves 110ms (from 150 ms to 40ms) from the logic to group torrents per tmdb and media type. It's not much compared to the rest of the request (~2s), but it's something. I could've got it down to 11 ms if laravel didn't have high overhead when accessing attributes, but it was 70ms before I started using ->getAttributeValue so that says something there.